move DownstreamData into downstream
