+++ miner-apps/translator/src/lib/sv1/downstream/downstream.rs
@@ -1,7 +1,7 @@
 use crate::{
     error::{self, TproxyError, TproxyErrorKind, TproxyResult},
     status::{handle_error, StatusSender},
-    sv1::downstream::data::DownstreamData,
+    sv1::downstream::SubmitShareWithChannelId,
 };
 use async_channel::{Receiver, Sender};
 use std::{
@@ -19,10 +19,11 @@ use stratum_apps::{
         sv1_api::{
             json_rpc::{self, Message},
             server_to_client,
+            utils::{Extranonce, HexU32Be},
         },
     },
     task_manager::TaskManager,
-    utils::types::{DownstreamId, Hashrate},
+    utils::types::{ChannelId, DownstreamId, Hashrate},
 };
 use tokio_util::sync::CancellationToken;
 use tracing::{debug, error, info, warn};
@@ -62,6 +63,81 @@ impl DownstreamChannelState {
         self.connection_token.cancel();
         self.downstream_sv1_receiver.close();
         self.downstream_sv1_sender.close();
+    }
+}
+
+#[derive(Debug)]
+pub struct DownstreamData {
+    pub channel_id: Option<ChannelId>,
+    pub extranonce1: Extranonce<'static>,
+    pub extranonce2_len: usize,
+    pub target: Target,
+    pub hashrate: Option<Hashrate>,
+    pub version_rolling_mask: Option<HexU32Be>,
+    pub version_rolling_min_bit: Option<HexU32Be>,
+    pub last_job_version_field: Option<u32>,
+    pub authorized_worker_name: String,
+    pub user_identity: String,
+    pub cached_set_difficulty: Option<json_rpc::Message>,
+    pub cached_notify: Option<json_rpc::Message>,
+    pub pending_target: Option<Target>,
+    pub pending_hashrate: Option<Hashrate>,
+    // Queue of Sv1 handshake messages received while waiting for SV2 channel to open
+    pub queued_sv1_handshake_messages: Vec<json_rpc::Message>,
+    // Stores pending shares to be sent to the sv1_server
+    pub pending_share: Option<SubmitShareWithChannelId>,
+    // Tracks the upstream target for this downstream, used for vardiff target comparison
+    pub upstream_target: Option<Target>,
+    // Timestamp of when the last job was received by this downstream, used for keepalive check
+    pub last_job_received_time: Option<Instant>,
+}
+
+impl DownstreamData {
+    pub fn new(hashrate: Option<Hashrate>, target: Target) -> Self {
+        DownstreamData {
+            channel_id: None,
+            extranonce1: vec![0; 8]
+                .try_into()
+                .expect("8-byte extranonce is always valid"),
+            extranonce2_len: 4,
+            target,
+            hashrate,
+            version_rolling_mask: None,
+            version_rolling_min_bit: None,
+            last_job_version_field: None,
+            authorized_worker_name: String::new(),
+            user_identity: String::new(),
+            cached_set_difficulty: None,
+            cached_notify: None,
+            pending_target: None,
+            pending_hashrate: None,
+            queued_sv1_handshake_messages: Vec::new(),
+            pending_share: None,
+            upstream_target: None,
+            last_job_received_time: None,
+        }
+    }
+
+    pub fn set_pending_target(&mut self, new_target: Target, downstream_id: DownstreamId) {
+        self.pending_target = Some(new_target);
+        debug!("Downstream {downstream_id}: Set pending target");
+    }
+
+    pub fn set_pending_hashrate(
+        &mut self,
+        new_hashrate: Option<Hashrate>,
+        downstream_id: DownstreamId,
+    ) {
+        self.pending_hashrate = new_hashrate;
+        debug!("Downstream {downstream_id}: Set pending hashrate");
+    }
+
+    pub fn set_upstream_target(&mut self, upstream_target: Target, downstream_id: DownstreamId) {
+        self.upstream_target = Some(upstream_target);
+        debug!(
+            "Downstream {downstream_id}: Set upstream target to {}",
+            upstream_target
+        );
     }
 }
