+++ miner-apps/translator/src/lib/sv1/downstream/mod.rs
@@ -1,4 +1,3 @@
-pub(super) mod data;
 pub mod downstream;
 
 use stratum_apps::{
